+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/detect/PreventOverwriteOfExternalizableObjectTest.java
@@ -54,16 +54,68 @@ class PreventOverwriteOfExternalizableObjectTest extends AbstractIntegrationTest
     }
 
     @Test
+    void testBadReadExternal6() {
+        performAnalysis("externalizable/BadExternalizableTest6.class");
+
+        assertNumOfBugs(2);
+        assertBug("BadExternalizableTest6", 22);
+        assertBug("BadExternalizableTest6", 23);
+    }
+
+    @Test
+    void testBadReadExternal7() {
+        performAnalysis("externalizable/BadExternalizableTest7.class");
+
+        assertNumOfBugs(2);
+        assertBug("BadExternalizableTest7", 22);
+        assertBug("BadExternalizableTest7", 23);
+    }
+
+    @Test
     void testGoodReadExternal() {
         performAnalysis("externalizable/GoodExternalizableTest.class");
-
         assertNumOfBugs(0);
     }
 
     @Test
     void testGoodReadExternal2() {
         performAnalysis("externalizable/GoodExternalizableTest2.class");
+        assertNumOfBugs(0);
+    }
 
+    @Test
+    void testGoodReadExternal3() {
+        performAnalysis("externalizable/GoodExternalizableTest3.class");
+        assertNumOfBugs(0);
+    }
+
+    @Test
+    void testGoodReadExternal4() {
+        performAnalysis("externalizable/GoodExternalizableTest4.class");
+        assertNumOfBugs(0);
+    }
+
+    @Test
+    void testGoodReadExternal5() {
+        performAnalysis("externalizable/GoodExternalizableTest5.class");
+        assertNumOfBugs(0);
+    }
+
+    @Test
+    void testGoodReadExternal6() {
+        performAnalysis("externalizable/GoodExternalizableTest6.class");
+        assertNumOfBugs(0);
+    }
+
+    @Test
+    void testGoodReadExternal7() {
+        performAnalysis("externalizable/GoodExternalizableTest7.class");
+        assertNumOfBugs(0);
+    }
+
+    @Test
+    void testGoodReadExternal8() {
+        performAnalysis("externalizable/GoodExternalizableTest8.class");
         assertNumOfBugs(0);
     }