+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/generic/GenericObjectType.java
@@ -213,13 +213,12 @@ public class GenericObjectType extends ObjectType {
      * @return the underlying ObjectType for this Generic Object
      */
     public ObjectType getObjectType() {
-        String cName = ClassName.fromFieldSignature(getSignature());
+        @DottedClassName
+        String cName = ClassName.fromFieldSignatureToDottedClassName(getSignature());
         if (cName == null) {
             throw new IllegalStateException("Can't provide ObjectType for " + this);
         }
-        @DottedClassName
-        String c = ClassName.toDottedClassName(cName);
-        return ObjectTypeFactory.getInstance(c);
+        return ObjectTypeFactory.getInstance(cName);
     }
 
     /**