+++ spotbugs/src/main/java/edu/umd/cs/findbugs/sarif/Invocation.java
@@ -13,18 +13,19 @@ import java.util.Objects;
  */
 class Invocation {
     private final int exitCode;
+
     @NonNull
-    private final String exitSignalName;
+    private final String exitCodeDescription;
     private final boolean executionSuccessful;
     @NonNull
     private final List<Notification> toolExecutionNotifications;
     @NonNull
     private final List<Notification> toolConfigurationNotifications;
 
-    Invocation(int exitCode, @NonNull String exitSignalName, boolean executionSuccessful, @NonNull List<Notification> toolExecutionNotifications,
+    Invocation(int exitCode, @NonNull String exitCodeDescription, boolean executionSuccessful, @NonNull List<Notification> toolExecutionNotifications,
             @NonNull List<Notification> toolConfigurationNotifications) {
         this.exitCode = exitCode;
-        this.exitSignalName = Objects.requireNonNull(exitSignalName);
+        this.exitCodeDescription = Objects.requireNonNull(exitCodeDescription);
         this.executionSuccessful = executionSuccessful;
         this.toolExecutionNotifications = Collections.unmodifiableList(toolExecutionNotifications);
         this.toolConfigurationNotifications = Collections.unmodifiableList(toolConfigurationNotifications);
@@ -34,7 +35,7 @@ class Invocation {
     JsonObject toJsonObject() {
         JsonObject result = new JsonObject();
         result.addProperty("exitCode", exitCode);
-        result.addProperty("exitSignalName", exitSignalName);
+        result.addProperty("exitCodeDescription", exitCodeDescription);
         result.addProperty("executionSuccessful", executionSuccessful);
 
         JsonArray execNotificationArray = new JsonArray();